// Code/UBind.PortalApp/src/app/models/details-list/details-list-item-type.enum.ts
export enum DetailsListItemGroupType {
    Phone = "phone",
    Email = "email",
    Person = "person",
    Company = "company",
    Relationships = "relationships",
    Dates = "dates",
    Others = "others",
    Website = "website",
    Messenger = "messenger",
    Social = "social",
    Address = "address",
    Status = "status",
    Details = "details",
    Properties = "properties",
    Account = 'account',
    Filters = "filters",
    Quotes = "quotes",
    Sort = "sort",
    Portal = "portal",
    LinkedIdentities = 'linkedIdentities',
    JsonSchema = 'jsonSchema'
}