// Code/UBind.PortalApp/src/app/models/details-list/details-list-item-card-type.enum.ts
export enum DetailsListItemCardType {
    ContactDetails = "Contact Details",
    Relationships = "Relationships",
    Dates = "Dates",
    Others = "Others",
    Details = "Details",
    Status = "Status",
    DataSources = "Data Sources",
    Template = "Template",
    Deployment = "Deployment",
    DeployedReleases = "Deployed Releases",
    AdditionalProperties = "Additional Properties",
    ApplicableDomainNames = "Applicable Domains Names",
    Account = 'Account',
    Claim = "Claim",
    Filters = "Filters",
    Quotes = "Quotes",
    Sort = "Sort",
    Optional = "Optional Data",
    Products = "Products",
    QuoteType = "Quote Type",
    Theme = 'Theme',
    SignInOptions = 'Sign In Options',
    SamlIdentityProvider = 'SAML Identity Provider',
    AutoProvisioning = 'Auto Provisioning',
    SamlAttributes = 'SAML Attributes',
    RoleMappings = 'Role Mappings',
    LinkedIdentities = 'Linked Identities',
    JsonSchema = 'JSON Schema',
    Usage = 'Usage',
}